A number of changes.
1) Modified test suites to make them more informative and check more
results (ie., check if a file has changed on disk as well as whether
the return code was correct.)

2) changed how IDL is run, now uses a direct spawn instead of a shell script

3) a number of updates to the lris throughput code to help with displaying the
output

+++ py/Planutil.py
@@ -3,7 +3,59 @@ from datetime import datetime
 from model import Plan, Frame, Instrument, Pipeline
 
 import re, os, os.path, glob, pyfits, tarfile, tempfile, sys, shutil
-import XIDLLongPlanutil,  Frameutil
+import Frameutil
+
+def dicttostr(dictionary):
+	if isinstance(dictionary,dict):
+		string= str(dictionary)
+	else:
+		# how did we get here?
+		string = str({})
+	return(string)
+
+def strtodict(string) :
+	if len(string):
+		dictionary = eval(string)
+	else:
+		dictionary = {}
+	return(dictionary)
+
+def inst_list():
+
+    return(["lrisred","lrisblue","kastb","kastr","nirspec"])
+
+def determine_nslit(frame,camera):
+    # XIDL and LRIS specific
+
+    minslit = 20 # default to something sensible.
+    if camera == "LRISBLUE":
+        minslit = 40 / frame.xbinning
+    elif camera == "LRISRED" or camera == "lrisred":
+        if 'DETECTOR' in frame.header.ascardlist().keys():
+            minslit = 40 / frame.xbinning
+        else:
+            minslit = 20 
+        # this only works for the old LRIS red
+        # this needs to be fixed for the new detectors.
+
+    return(minslit)
+
+def linkfiles(frame,plan,datapath):
+    """linkfiles(plan,plan,datapath)
+    This deals with the fact that XIDL can only have one datapath.
+    Because input data can come from multiple sources, we will make links.
+    This will assume a single frame, a plan and a datapath.
+    
+    """
+    msg = ''
+    inpath = os.path.join(datapath,frame.path,frame.display_name)
+    outpath = os.path.join(datapath,plan.finalpath)
+    if os.path.isfile(inpath) and os.path.isdir(outpath) and not os.path.isfile(os.path.join(outpath,frame.display_name)):
+        try:
+            os.link(inpath,os.path.join(outpath,frame.display_name))
+        except IOError as e:
+            msg = "%s" % e.strerror
+    return(msg)
 
 
 def makeplan_displayname(frame):
@@ -68,10 +120,11 @@ def buildplan(frame,planname,datapath,pipeline,flag):
         files = glob.glob(fullpath + "/*.???")
         for cfile in files:
             os.remove(os.path.join(cfile))
-        files = glob.glob(fullpath + "/Science/*")
-        for cfile in files:
-            os.remove(os.path.join(cfile))
-        os.rmdir(os.path.join(fullpath,"Science"))
+        if os.path.isdir(os.path.join(fullpath,"Science")):
+            files = glob.glob(fullpath + "/Science/*")
+            for cfile in files:
+                os.remove(os.path.join(cfile))
+            os.rmdir(os.path.join(fullpath,"Science"))
         
     plan.instrument = frame.instrument
 
@@ -79,23 +132,121 @@ def buildplan(frame,planname,datapath,pipeline,flag):
        
     return (plan)
 
-def updateplandata(plan,frame,datapath):
+def updateplandata(plan,datapath):
 
     """This generates the pipeline specific information.
+    This will ONLY work for XIDL and LRIS now. This does one thing.
+    It builds up the actual plan file content, the plan file is written later.
+
     """
 
-    plan = XIDLLongPlanutil.buildplanlong(plan,datapath,plan.instrument.name )
+#    plan = XIDLLongPlanutil.buildplanlong(plan,datapath,plan.instrument.name )
+
+    flags = strtodict(plan.runstr)
+
+
+    plan.data = "logfile, '" + plan.name + ".log'   # Log file\n"
+    plan.data += "plotfile, '" + plan.name + ".ps'   # Plot file\n"
+    plan.data += "indir \'" + os.path.abspath(os.path.join(datapath,plan.finalpath)) + "/\'\n"
+    plan.data += "tempdir Temp\nscidir  Science\n"
+    if flags.has_key('maxobj'):
+        try:
+            plan.data += "maxobj %d \n" %  int(flags['maxobj'])
+        except ValueError:
+            plan.data += "maxobj %d \n" %  5
+
+    else:
+        plan.data += "maxobj %d \n" %  plan.maxobj
+    if flags.has_key('minslit'):
+        try:
+            plan.data += "minslit %f \n" % float(flags['minslit'])
+        except ValueError:
+            if len(plan.frames) > 0:
+                slitsize = determine_nslit(plan.frames[0],plan.instrument.name)
+            else:
+                slitsize = 20
+            plan.data += "minslit %f \n" % slitsize
+    else:
+            if len(plan.frames) > 0:
+                slitsize = determine_nslit(plan.frames[0],plan.instrument.name)
+            else:
+                slitsize = plan.minslit
+            plan.data += "minslit %f \n" % slitsize
+
+    if flags.has_key('reduxthresh'):
+        try:
+            plan.data += "reduxthresh %f \n" % float(flags['reduxthresh'])
+        except ValueError:
+            plan.data += "reduxthresh %f \n" % 0.01
+    else:
+        plan.data += "reduxthresh %f \n" % plan.reduxthresh
+
+    plan.data += "idlutilsVersion \'NOCVS:idlutils\'\n"
+    plan.data += "LongslitVersion \'NOCVS:Unknown\'\n"
+
+    plan.data += "\ntypedef struct {\n char filename[45];\n"
+    plan.data += " char flavor[8];\n char target[12];\n"
+    plan.data += " float exptime;\n char instrument[9];\n"
+    plan.data += " char grating[9];\n" 
+    if plan.instrument.name == "LRISRED":
+        plan.data += " char wave[10];\n"
+    else:
+        plan.data += " char wave[1];\n"
+    plan.data += " char maskname[18];\n} LEXP;\n"
 
     return(plan)
 
+def determine_numkind_flats(plan):
+
+    numflat = { "Flat" : 0,
+                "DmFlat" : 0,
+                "IntFlat" : 0,
+                "PixFlat" : 0
+                }
+                
+    for frame in plan.frames:
+        if frame.type in numflat:
+            numflat[frame.type] += 1
+
+    return(numflat)
     
 def writeplan(plan,datapath):
 
     """Writes the plan file.  This is what the software will actually execute.
+    This is hard coded for the XIDL/LRIS specific content.
     """
 
 #    cleanoutplandir(plan,datapath)
-    XIDLLongPlanutil.writelongplan(plan,datapath)
+#    XIDLLongPlanutil.writelongplan(plan,datapath)
+# def writelongplan(plan,datapath):
+
+    camera = plan.instrument.name
+    planpath = os.path.join(datapath, plan.finalpath, "plan.par")
+
+    if os.path.exists(planpath):
+        os.remove(planpath)
+
+    numkindflat = determine_numkind_flats(plan)
+    for frame in plan.frames:
+        msg =linkfiles(frame,plan,datapath)
+        if msg:
+            sys.exit(msg)
+            return 
+    framestr = buildXIDLframestr(buildXIDLframelist(plan,camera))
+
+    planfile = open(planpath,'w')
+    if plan.data == '':
+        updateplandata(plan,datapath)
+
+    # if camera == "lrisblue" and numkindflat['PixFlat'] > 0:
+    #     writepixflatplan(plan,datapath,camera)
+    # elif camera == "kastb" and numkindflat['PixFlat'] > 0:
+    #     writepixflatplan(plan,datapath,camera)
+
+    planfile.write(plan.data)
+    planfile.write(framestr)
+    planfile.close()
+    return()
 
 
 def gatherframelist(plan):
@@ -107,7 +258,7 @@ def gatherframelist(plan):
 
     framelist = ""
 
-    framelist = XIDLLongPlanutil.buildframelist(plan,plan.instrument.name )
+    framelist = buildXIDLframelist(plan,plan.instrument.name )
 
     return(framelist)
 
@@ -133,13 +284,71 @@ def buildgenericframelist(plan):
         
     return(framelist)
 
-def buildframestr(framelist):
+def buildXIDLframestr(framelist):
 
     framestr = ""
     for framestuff in framelist:
         framestr += "\n" + framestuff
 
     return(framestr)
+
+def genXIDLframestr(frame,camera):
+
+    outtag = { "Object" : "\"science\"",
+               "IntFlat" : "\"iflat\"",
+               "Flat" : "\"twiflat\"",
+               "DmFlat" : "\"domeflat\"",
+               "Line" : "\"arc\"",
+               "PixFlat" : "\"domeflat\"",
+               "Trace" : "\"std\""
+               }
+
+    framestr = "LEXP " + frame.display_name
+    if camera == "lrisred" and (frame.type== "Flat" or frame.type=="DmFlat" or frame.type=="IntFlat"):
+        framestr  += " %s " % "bothflat"
+    elif camera == "lrisblue" and (frame.type== "Flat" or frame.type=="DmFlat" or frame.type=="IntFlat"):
+#        framestr  += " %s " % '\"twiflat\"'
+        framestr  += " %s " % '\"bothflat\"'
+    elif camera == "kastr" and (frame.type== "Flat" or frame.type=="DmFlat" or frame.type=="IntFlat"):
+        framestr  += " %s " % "bothflat"
+    elif camera == "kastb" and (frame.type== "Flat" or frame.type=="DmFlat" or frame.type=="IntFlat"):
+        framestr  += " %s " % '\"twiflat\"'
+    else:
+        framestr  += " %s " % outtag[frame.type]
+
+    plantarget = frame.target
+    planaperture = frame.aperture
+    if frame.target:
+        framestr += " \"%s\" " % plantarget 
+    else:
+        framestr += " \"%s\" " % planaperture
+    framestr += " %.1f " % float(frame.exptime)
+    framestr += " %s " % camera 
+    framestr += " %s " % frame.grating 
+    if frame.wavelength :
+        framestr += " %.2f " % float(frame.wavelength)
+    else :
+        framestr += " \"\" "
+    framestr += " \"%s\" \n" % planaperture
+    return(framestr)
+
+
+def buildXIDLframelist(plan,camera):
+
+    framelist = []
+
+
+    numflat = determine_numkind_flats(plan)
+
+    for frametype in plan.pipeline.typelist():
+
+        frames = plan.framelisttype(type=frametype)
+
+        for frame in frames:
+            framestr = genXIDLframestr(frame,camera)
+            framelist.append(framestr)
+
+    return(framelist)
 
 
 def writegenericplan(plan,datapath):
@@ -159,21 +368,70 @@ def writegenericplan(plan,datapath):
     plan.data += "\n"
 
     planfile.write(plan.data)
-    planfile.write(buildframestr(buildgenericframelist(plan)))
+    planfile.write(buildXIDLframestr(buildgenericframelist(plan)))
     planfile.close()
     return()
 
 def genrunstr(plan,datapath):
 
     finalrunstr = ""
-    if plan.instrument.name in XIDLLongPlanutil.inst_list():
-        finalrunstr = XIDLLongPlanutil.modrunstr(plan,datapath)
+    if plan.instrument.name in inst_list():
+        finalrunstr = XIDLmodrunstr(plan,datapath)
     else:
         finalrunstr = plan.pipeline.runstr % plan.display_name
         finalrunstr = finalrunstr + '\n'
 
     return(finalrunstr)
 
+def XIDLmodrunstr(plan,datapath):
+
+    # This relies on the run string 
+
+    # The actual item stored in the model as the runstr is stripped of leading and trailing spaces.
+    # If the camera is the lris blue and there are not pixel flats, then the PIXFLAT_ARCHIVE flag is added.
+    
+    # In addition, if the user has taken pixel flats (images with the dispersing element but without an aperture, usually on the sky)
+    # then we need to run the pipeline twice.  Once with just the pixel flats in their own plan file, and once with everything.
+
+    flags = strtodict(plan.runstr)
+
+
+    runstr = plan.pipeline.runstr # % plan.display_name
+    runstr = runstr + '\n'
+
+
+    numflat = determine_numkind_flats(plan)
+    if numflat["PixFlat"] == 0 and plan.instrument.name == "lrisblue":
+        flags["PIXFLAT_ARCHIVE"] = "On"
+
+    flagstr = ""
+    if isinstance(flags,dict):
+        for flagkey in flags.keys():
+            flag = ""
+            if flags[flagkey] == "On" and flagkey != "skytrace":
+                flag = ",/" + flagkey.upper()
+            elif flagkey == "skytrace" and flags[flagkey] == "On":
+                flag = "," + flagkey.upper() + "=1"
+            elif flagkey == "skytrace" and flags[flagkey] == "Off":
+                flag = "," + flagkey.upper() + "=0"
+
+            if flagkey == "slitdefs":
+                flag = ",ADD_SLITS='slitlocations'"
+
+                slitpath = os.path.join(datapath,plan.finalpath,'slitlocations')
+                slitfile = open(slitpath,'w')
+                slitfile.write(str(flags['slitdefs']))
+                slitfile.close()
+        
+            flagstr += flag
+        flagstr = " ".join([flagstr,'"','|'])
+        
+        runstr = re.sub('\"\s\|',flagstr,runstr)
+
+    plan.runstr = dicttostr(flags)
+
+
+    return(runstr)
 
 def writerunstr(runpath,datapath,finalrunstr,idlenv):
     """Writes the actual run string.  This what gets executed by the runjob.py agent
@@ -184,7 +442,7 @@ def writerunstr(runpath,datapath,finalrunstr,idlenv):
         
     # FIX
     envstuff = u'#!/bin/csh\nsource ' + idlenv + '\n'
-# 'setenv IDL_DIR /usr/local/idl_80/idl\nsource /home/lris/idl/IPM/THROUGHPUT/watchdir_env/idlenv\n'
+    # 'setenv IDL_DIR /usr/local/idl_80/idl\nsource /home/lris/idl/IPM/THROUGHPUT/watchdir_env/idlenv\n'
     
     rawpath = os.path.join(datapath)
     #    envstuff = envstuff + "#setenv DEIMOS_DATA " + rawpath + "\n"
